move config generation logic back to l3Clos.py
