+++ jnpr/openclos/conf/closTemplate.yaml
@@ -25,3 +25,23 @@ pods:
         # if not placed under this dir, the file would not be accessible from http server
         # and ZTP process will be broken, it overrides ztp.junosImage setting
         junosImage : jinstall-qfx-5-flex-13.2X51-D21.2-domestic-signed.tgz
+    anotherPod:
+        spineCount : 4
+        # possible options for leafDeviceType are QFX5100-24Q
+        spineDeviceType : QFX5100-24Q
+        leafCount : 6
+        # possible options for leafDeviceType are QFX5100-96S, QFX5100-48S
+        leafDeviceType : QFX5100-48S
+        hostOrVmCountPerLeaf : 254
+        interConnectPrefix : 192.168.0.0
+        vlanPrefix : 172.16.0.0
+        loopbackPrefix : 10.0.0.0
+        spineAS : 100
+        leafAS : 200
+        # possible options for topologyType are threeStage, fiveStageRealEstate, fiveStagePerformance
+        topologyType : threeStage
+        inventory : inventoryLabKurt.json
+        # the image file should be placed under <install dir>/jnpr/openclos/conf/junosImages
+        # if not placed under this dir, the file would not be accessible from http server
+        # and ZTP process will be broken, it overrides ztp.junosImage setting
+        junosImage : jinstall-qfx-5-flex-13.2X51-D21.2-domestic-signed.tgz
